requirements.txt
Django==3.1.7
django-cors-headers
django-graphql-jwt
graphene-django
graphene>=2.0